# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.SendMessageRequest:
    description: Request payload for sending WhatsApp messages
    properties:
      caption:
        description: Caption for media messages)
        example: Check this out
        type: string
      file_name:
        description: File name (for document messages)
        example: document.pdf
        type: string
      jid:
        description: Target chat JID (also accept "jid" from frontend)
        example: 6281234567890@s.whatsapp.net
        type: string
      media_url:
        description: Media URL (for media messages)
        example: https://example.com/image.jpg
        type: string
      message:
        description: Text content (also accept "message" from frontend)
        example: Hello World
        type: string
      message_type:
        description: Message type
        enum:
        - text
        - image
        - video
        - audio
        - document
        example: text
        type: string
      quoted_message_id:
        description: Quoted message ID for replies
        example: 3EB0ABCD1234567890
        type: string
      session_id:
        description: Session ID
        example: 019b6954-9536-7d55-a444-45acf931f60f
        type: string
    required:
    - jid
    - message_type
    - session_id
    type: object
...
  dto.WebhookConfigRequest:
    properties:
      enabled:
        type: boolean
      headers:
        additionalProperties:
          type: string
        type: object
      max_retries:
        maximum: 10
        minimum: 0
        type: integer
      retry_delay_seconds:
        maximum: 300
        minimum: 1
        type: integer
      secret:
        type: string
      url:
        type: string
    required:
    - url
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  response.WhatsAppBaseResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API server untuk mengelola WhatsApp messaging, status, dan webhook
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: WhatsApp API Server
  version: "1.0"
paths:
...
  /whatsapp/send:
    post:
      consumes:
      - application/json
      description: Send text, image, document, or other media messages via WhatsApp
      parameters:
      - description: Message data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/dto.SendMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.WhatsAppBaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Send WhatsApp message
      tags:
      - WhatsApp
...
  /whatsapp/webhook/configure:
    post:
      consumes:
      - application/json
      description: Configure webhook URL and retry settings for WhatsApp events
      parameters:
      - description: Webhook configuration
        in: body
        name: webhook
        required: true
        schema:
          $ref: '#/definitions/dto.WebhookConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.WhatsAppBaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Configure webhook
      tags:
      - WhatsApp Webhook
...
  /whatsapp/webhook/test:
    post:
      consumes:
      - application/json
      description: Send a test webhook to verify configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.WhatsAppBaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Test webhook
      tags:
      - WhatsApp Webhook
schemes:
- http
- https
swagger: "2.0"